requirements.txt
openai-whisper
soundcard
websocket-client
websockets
numpy
